fix(financeiro): hotfix produção — COUNT(*) bigint → cast ::int nos mapeamentos Dapper (comissões + paginação do extrato)
Briefing: planejamentos/2026-06-11_001_financeiro-f7-consolidacao.md

Causa: Dapper + Npgsql 10 — QueryMultipleAsync ReadSingleAsync<int>() e
QueryAsync<T> com campo int não convertem Int64 via Convert.ChangeType
nesse path; lançam InvalidCastException → 500 em produção.

3 ocorrências corrigidas em ConsolidacaoFinanceiraQueryRepository:
1. COUNT(*) de paginação do ListarExtrato (sqlBase) → quebrava todo GET /api/financeiro/extrato com dados
2. COUNT(*) AS Atendimentos na query sqlConsultaProcedimento (comissões) → 500 na aba Comissões
3. COUNT(*) AS Atendimentos na query sqlCirurgia (comissões) → idem

Varredura completa em todos os repositórios Dapper F1-F7: demais COUNT(*)
sem ::int usam ExecuteScalarAsync<int> (converte via boxing) — sem risco
imediato. Risco latente em RelatorioQueryRepository (AgendamentosPorStatus,
AgendamentosPorDia, FaturamentoCategoria) documentado como Tipo A separado.

+++ backend/src/Services/Imedto.Backend.Infrastructure/Database/Repositories/ConsolidacaoFinanceiraQueryRepository.cs
@@ -119,7 +119,7 @@ public class ConsolidacaoFinanceiraQueryRepository
             """;
 
         var sql = $"""
-            SELECT COUNT(*) {sqlBase};
+            SELECT COUNT(*)::int {sqlBase};
 
             SELECT
                 l.id                                                  AS Id,
@@ -283,7 +283,7 @@ public class ConsolidacaoFinanceiraQueryRepository
                 cc.profissional_id                                                      AS ProfissionalUsuarioId,
                 COALESCE(u.nome_completo, u.email, cc.profissional_id::text)            AS Nome,
                 v.especialidade_convidada                                               AS Especialidade,
-                COUNT(*)                                                                AS Atendimentos,
+                COUNT(*)::int                                                            AS Atendimentos,
                 SUM(pp.valor_liquido)                                                   AS Faturamento,
                 COALESCE(cfg.percentual, @PercentualPadrao)                             AS PercentualConfig,
                 ROUND(SUM(pp.valor_liquido) * COALESCE(cfg.percentual, @PercentualPadrao) / 100, 2) AS Comissao,
@@ -332,7 +332,7 @@ public class ConsolidacaoFinanceiraQueryRepository
                 oe.profissional_usuario_id                                      AS ProfissionalUsuarioId,
                 COALESCE(u.nome_completo, u.email, oe.profissional_usuario_id::text) AS Nome,
                 v.especialidade_convidada                                        AS Especialidade,
-                COUNT(*)                                                         AS Atendimentos,
+                COUNT(*)::int                                                     AS Atendimentos,
                 ROUND(SUM(oe.valor * CASE WHEN pc.total_cobrado > 0
                     THEN pc.total_pago / pc.total_cobrado ELSE 1 END), 2)       AS Faturamento,
                 NULL::numeric                                                    AS PercentualConfig,
